feat: transform lang selector to list of buttons
