+++ src/components/language-selector/language-selector.module.scss
@@ -1,0 +1,26 @@
+.language-selector {
+  padding: 0;
+  margin: 0;
+  list-style-type: none;
+
+  li {
+    display: inline-block;
+    margin: 0;
+  }
+
+  button {
+    margin-right: 10px;
+    padding: 0;
+    border: 0;
+    text-decoration: underline;
+    cursor: pointer;
+    color: inherit;
+
+    &:disabled {
+      text-decoration: none;
+      color: inherit;
+      font-weight: bold;
+      cursor: default;
+    }
+  }
+}
